# apigateway/docs/swagger.yaml
definitions:
  requests.AuthRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  requests.CreateResetPasswordRequest:
    properties:
      confirm_password:
        minLength: 6
        type: string
      password:
        minLength: 6
        type: string
      reset_token:
        type: string
    required:
    - confirm_password
    - password
    - reset_token
    type: object
...
  requests.CreateUserRequest:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - confirm_password
    - email
    - firstname
    - lastname
    - password
    type: object
  requests.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  requests.RefreshTokenRequest:
    properties:
      refresh_token:
        minLength: 1
        type: string
    required:
    - refresh_token
    type: object
...
  response.ApiResponseForgotPassword:
    properties:
      messsage:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseGetMe:
    properties:
      data:
        $ref: '#/definitions/response.UserResponse'
      messsage:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseLogin:
    properties:
      data:
        $ref: '#/definitions/response.TokenResponse'
      messsage:
        type: string
      status:
        type: string
    type: object
...
  response.ApiResponseRefreshToken:
    properties:
      data:
        $ref: '#/definitions/response.TokenResponse'
      messsage:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseRegister:
    properties:
      data:
        $ref: '#/definitions/response.UserResponse'
      messsage:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseResetPassword:
    properties:
      messsage:
        type: string
      status:
        type: string
    type: object
...
  response.ApiResponseVerifyCode:
    properties:
      messsage:
        type: string
      status:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
  response.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  response.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      lastname:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user using the provided email and password.
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.ApiResponseLogin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Authenticate a user
      tags:
      - Auth
  /api/auth/me:
    get:
      description: Retrieves the current user's information using a valid access token
        from the Authorization header.
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.ApiResponseGetMe'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      - BearerToken: []
      summary: Get current user information
      tags:
      - Auth
  /api/auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Refreshes the access token using a valid refresh token.
      parameters:
      - description: Refresh token data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.ApiResponseRefreshToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Refresh access token
      tags:
      - Auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user with the provided details.
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.ApiResponseRegister'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Initiates password reset by sending a reset token to the provided
        email.
      parameters:
      - description: Forgot Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseForgotPassword'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Sends a reset token to the user's email
      tags:
      - Auth
...
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Allows user to reset their password using a valid reset token.
      parameters:
      - description: Reset Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseResetPassword'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Resets the user's password using a reset token
      tags:
      - Auth
  /auth/verify-code:
    get:
      description: Verifies the user's email using the verification code provided
        in the query parameter.
      parameters:
      - description: Verification Code
        in: query
        name: verify_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseVerifyCode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Verifies the user using a verification code
      tags:
      - Auth
swagger: "2.0"